+++ Integrations/Active_Directory_Query/CHANGELOG.md
@@ -1,4 +1,13 @@
 ## [Unreleased]
+ - Added 2 commands.
+   - ***ad-create-group*** 
+   - ***ad-delete-group***
+
+
+## [19.12.1] - 2019-12-25
+Fixed a typo in the name of the **custom-field-data** argument.
+
+## [19.12.0] - 2019-12-10
 Improved handling of error messages.
 
 ## [19.11.0] - 2019-11-12