+++ Integrations/Slack/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
 
+## [19.12.0] - 2019-12-10
+  - Fixed an issue where mirrored investigations contained mismatched user names.
+  - Added reporter and reporter email as labels to incidents that are created by direct messages.
+
 ## [19.11.1] - 2019-11-26
 Added Slack API rate limit call handling.
 Added an optional parameter to specify a proxy URL to use with the Slack API.